requirements.txt
aiogram==3.17.0
apscheduler==3.11.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.37
asyncpg==0.30.0
lxml==5.3.0